Reworked Interactables
Created a new class that can be easily subclassed in the form of InteractableObject. This has a customizable Toggle() function that can be used to create different interactions for different types of objects (eg. Doors, Switches).
SwitchHandler may contain some unnecessary redundancy as it calls the Toggle() of the remote door. But ObjectIsTriggerable could possibly meet the same requirement. Would love a second set of eyes on that issue.